[REFACTOR/SERVER] move stuff to client_management
from fpsgame/server.cpp
